+++ src/main/java/dinkplugin/DinkPluginConfig.java
@@ -2065,7 +2065,7 @@ public interface DinkPluginConfig extends Config {
         section = externalSection
     )
     default boolean notifyExternal() {
-        return true; // enabled by default, unlike other notifiers
+        return false;
     }
 
     @ConfigItem(